+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/externalevents/AlarmReceiver.java
@@ -17,16 +17,11 @@
 package nodomain.freeyourgadget.gadgetbridge.externalevents;
 
 
-import android.Manifest;
 import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.location.Criteria;
-import android.location.Location;
-import android.location.LocationManager;
 
 import net.e175.klaus.solarpositioning.DeltaT;
 import net.e175.klaus.solarpositioning.SPA;
@@ -38,11 +33,10 @@ import java.util.Calendar;
 import java.util.GregorianCalendar;
 import java.util.TimeZone;
 
-import androidx.core.app.ActivityCompat;
 import nodomain.freeyourgadget.gadgetbridge.BuildConfig;
 import nodomain.freeyourgadget.gadgetbridge.GBApplication;
 import nodomain.freeyourgadget.gadgetbridge.model.CalendarEventSpec;
-import nodomain.freeyourgadget.gadgetbridge.util.Prefs;
+import nodomain.freeyourgadget.gadgetbridge.util.GBPrefs;
 
 public class AlarmReceiver extends BroadcastReceiver {
     private static final Logger LOG = LoggerFactory.getLogger(AlarmReceiver.class);
@@ -86,26 +80,11 @@ public class AlarmReceiver extends BroadcastReceiver {
         GBApplication.deviceService().onDeleteCalendarEvent(CalendarEventSpec.TYPE_SUNRISE, id_tomorrow);
         GBApplication.deviceService().onDeleteCalendarEvent(CalendarEventSpec.TYPE_SUNSET, id_tomorrow);
 
-        Prefs prefs = GBApplication.getPrefs();
+        GBPrefs gbPrefs = GBApplication.getGBPrefs();
+        float[] longlat = gbPrefs.getLongLat(context);
+        float longitude = longlat[0];
+        float latitude = longlat[1];
 
-        float latitude = prefs.getFloat("location_latitude", 0);
-        float longitude = prefs.getFloat("location_longitude", 0);
-        LOG.info("got longitude/latitude from preferences: " + latitude + "/" + longitude);
-
-        if (ActivityCompat.checkSelfPermission(context, Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED &&
-                prefs.getBoolean("use_updated_location_if_available", false)) {
-            LocationManager locationManager = (LocationManager) context.getSystemService(Context.LOCATION_SERVICE);
-            Criteria criteria = new Criteria();
-            String provider = locationManager.getBestProvider(criteria, false);
-            if (provider != null) {
-                Location lastKnownLocation = locationManager.getLastKnownLocation(provider);
-                if (lastKnownLocation != null) {
-                    latitude = (float) lastKnownLocation.getLatitude();
-                    longitude = (float) lastKnownLocation.getLongitude();
-                    LOG.info("got longitude/latitude from last known location: " + latitude + "/" + longitude);
-                }
-            }
-        }
         GregorianCalendar[] sunriseTransitSetTomorrow = SPA.calculateSunriseTransitSet(dateTimeTomorrow, latitude, longitude, DeltaT.estimate(dateTimeTomorrow));
 
         CalendarEventSpec calendarEventSpec = new CalendarEventSpec();