+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Changelog
 
 #### 0.47.2 (WIP)
+* Amazfit Bip S: Send sunrise and sunset on latest firmware if enabled
 * Huami: Support new firmware update protocol (fixes firmware flashing with firmware 2.1.1.50/4.1.5.55 on Amazfit Bip S)
 * Huami: Allow flashing latest GPS firmware
 * InfiniTime: Add support for music control